backend/Arthaniti_ai_backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
python-dotenv
numpy
pandas
matplotlib
seaborn
jinja2
requests
httpx
google-generativeai     # Gemini
python-multipart
loguru
